--- log/log.go
@@ -1,4 +1,0 @@
-package log
-
-// Trace defines if we want to launch the enforcer in Tracing mode (Print all packets)
-var Trace = false
