+++ supervisor/iptablesctrl/iptables_test.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/aporeto-inc/trireme-lib/constants"
 	"github.com/aporeto-inc/trireme-lib/enforcer/utils/fqconfig"
+	"github.com/aporeto-inc/trireme-lib/internal/portset"
 	"github.com/aporeto-inc/trireme-lib/policy"
-	"github.com/aporeto-inc/trireme-lib/portset"
 	"github.com/aporeto-inc/trireme-lib/supervisor/provider"
 	. "github.com/smartystreets/goconvey/convey"
 )
