+++ supervisor/supervisor.go
@@ -12,8 +12,8 @@ import (
 	"github.com/aporeto-inc/trireme-lib/constants"
 	"github.com/aporeto-inc/trireme-lib/enforcer/policyenforcer"
 	"github.com/aporeto-inc/trireme-lib/enforcer/utils/fqconfig"
+	"github.com/aporeto-inc/trireme-lib/internal/portset"
 	"github.com/aporeto-inc/trireme-lib/policy"
-	"github.com/aporeto-inc/trireme-lib/portset"
 	"github.com/aporeto-inc/trireme-lib/supervisor/ipsetctrl"
 	"github.com/aporeto-inc/trireme-lib/supervisor/iptablesctrl"
 )
